+++ sickrage/core/databases/main/__init__.py
@@ -24,7 +24,8 @@ import sickrage
 from sickrage.core.databases import srDatabase
 from sickrage.core.databases.main.index import MainTVShowsIndex, MainTVEpisodesIndex, MainIMDBInfoIndex, \
     MainXEMRefreshIndex, MainSceneNumberingIndex, MainIndexerMappingIndex, MainHistoryIndex, \
-    MainBlacklistIndex, MainWhitelistIndex, MainFailedSnatchHistoryIndex, MainFailedSnatchesIndex, MainVersionIndex
+    MainBlacklistIndex, MainWhitelistIndex, MainFailedSnatchHistoryIndex, MainFailedSnatchesIndex, MainVersionIndex, \
+    MainTVEpisodesByIndexerIDIndex
 
 
 class MainDB(srDatabase):
@@ -34,6 +35,7 @@ class MainDB(srDatabase):
         'version': MainVersionIndex,
         'tv_shows': MainTVShowsIndex,
         'tv_episodes': MainTVEpisodesIndex,
+        'tv_episodes_by_indexerid': MainTVEpisodesByIndexerIDIndex,
         'imdb_info': MainIMDBInfoIndex,
         'xem_refresh': MainXEMRefreshIndex,
         'scene_numbering': MainSceneNumberingIndex,